requirements.txt
TotalSegmentator==2.11.0
SimpleITK==2.4.0
nibabel==5.3.2
numpy==2.0.2
pandas==2.2.3
matplotlib==3.10.1
PyYAML==6.0.2
